Include stop sequences in output frames
Each NAL unit is separated by a stop sequence of 00000001 or similar.
This was not being included in the output data.

It seems that esparser or the decoder expects this, otherwise motion
is not rendered correctly.

Now we send the stop sequence at the start of each output buffer
rather than skipping over it.